SmartContract/flask-api/requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
eth-utils==2.3.1  # để keccak
python-dotenv==1.0.1
requests==2.32.3

